+++ packages/spacecat-shared-content-client/CHANGELOG.md
@@ -1,3 +1,38 @@
+# [@adobe/spacecat-shared-content-client-v1.2.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.1.22...@adobe/spacecat-shared-content-client-v1.2.0) (2024-11-12)
+
+
+### Features
+
+* create content client directly from domain ([17aec10](https://github.com/adobe/spacecat-shared/commit/17aec10cabb7b50d35588054587a96032b52ef73))
+
+# [@adobe/spacecat-shared-content-client-v1.1.22](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.1.21...@adobe/spacecat-shared-content-client-v1.1.22) (2024-11-11)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-helix-content-sdk to v1.2.7 ([#430](https://github.com/adobe/spacecat-shared/issues/430)) ([a7de0ee](https://github.com/adobe/spacecat-shared/commit/a7de0eeda9eda2f3460deb4df3a12b2894fba455))
+
+# [@adobe/spacecat-shared-content-client-v1.1.21](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.1.20...@adobe/spacecat-shared-content-client-v1.1.21) (2024-11-11)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#434](https://github.com/adobe/spacecat-shared/issues/434)) ([b71b615](https://github.com/adobe/spacecat-shared/commit/b71b61528513821f9e34c50a095d47cb4e14f8db))
+
+# [@adobe/spacecat-shared-content-client-v1.1.20](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.1.19...@adobe/spacecat-shared-content-client-v1.1.20) (2024-11-08)
+
+
+### Bug Fixes
+
+* engine spec ([#433](https://github.com/adobe/spacecat-shared/issues/433)) ([19dd309](https://github.com/adobe/spacecat-shared/commit/19dd30956c0a9d35ea343c580e589205bfdbdfd8))
+
+# [@adobe/spacecat-shared-content-client-v1.1.19](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.1.18...@adobe/spacecat-shared-content-client-v1.1.19) (2024-11-08)
+
+
+### Bug Fixes
+
+* engine spec ([#431](https://github.com/adobe/spacecat-shared/issues/431)) ([f19600e](https://github.com/adobe/spacecat-shared/commit/f19600e74ae7ee42a97c08d5b1c30db79c8ec13d))
+
 # [@adobe/spacecat-shared-content-client-v1.1.18](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.1.17...@adobe/spacecat-shared-content-client-v1.1.18) (2024-11-07)